+++ npm-packages/react-api/src/hooks/use-agent-tools.ts
@@ -1,13 +1,18 @@
 import { AgentTool, IMCMessage, IMCMessageTypeEnum } from "@pulse-editor/shared-utils";
 import useIMC from "../lib/hooks/use-imc";
 
-export default function useAgentTools(moduleName: string) {
+/**
+ * Add or use agent tools in the editor.
+ * @param moduleName 
+ * @returns 
+ */
+export default function useAgentTools() {
   const receiverHandlerMap = new Map<
     IMCMessageTypeEnum,
     (senderWindow: Window, message: IMCMessage) => Promise<void>
   >();
 
-  const { imc } = useIMC(moduleName, receiverHandlerMap);
+  const { imc } = useIMC(receiverHandlerMap);
 
   async function installAgentTool(tool: AgentTool) {
     if (!imc) {