+++ web/app/layout.tsx
@@ -1,11 +1,10 @@
 import type { Metadata } from "next";
 import "./globals.css";
-import { WrappedHeroUIProvider } from "../components/providers/wrapped-hero-ui-provider";
+import { WrappedHeroUIProvider } from "@/components/providers/wrapped-hero-ui-provider";
 import EditorContextProvider from "@/components/providers/editor-context-provider";
 import { Toaster } from "react-hot-toast";
 import "material-icons/iconfont/material-icons.css";
 import CapacitorProvider from "@/components/providers/capacitor-provider";
-import Nav from "@/components/interface/nav";
 import RemoteExtensionProvider from "@/components/providers/remote-extension-provider";
 
 export const metadata: Metadata = {
@@ -26,7 +25,7 @@ export default function RootLayout({
             <EditorContextProvider>
               <RemoteExtensionProvider>
                 <Toaster />
-                <Nav>{children}</Nav>
+                {children}
               </RemoteExtensionProvider>
             </EditorContextProvider>
           </WrappedHeroUIProvider>