+++ web/components/providers/editor-context-provider.tsx
@@ -32,6 +32,7 @@ const defaultEditorStates: EditorStates = {
   explorerSelectedNodeRefs: [],
   pressedKeys: [],
   openedViewModels: [],
+  viewIds: [],
 };
 
 export default function EditorContextProvider({