+++ web/components/views/loaders/file-view-loader.tsx
@@ -1,9 +1,14 @@
-import { Extension, FileViewModel, InstalledAgent } from "@/lib/types";
+import { Extension, InstalledAgent } from "@/lib/types";
 import { useContext, useEffect, useState } from "react";
 import { EditorContext } from "../../providers/editor-context-provider";
 import FileViewLayout from "../layout";
 import ExtensionLoader from "../../misc/extension-loader";
-import { Agent, IMCMessage, IMCMessageTypeEnum } from "@pulse-editor/shared-utils";
+import {
+  Agent,
+  FileViewModel,
+  IMCMessage,
+  IMCMessageTypeEnum,
+} from "@pulse-editor/shared-utils";
 import Loading from "../../interface/loading";
 import useAgentRunner from "@/lib/hooks/use-agent-runner";
 import { useTheme } from "next-themes";
@@ -69,7 +74,7 @@ export default function FileViewLoader({
     }
 
     if (fileUri) {
-      // Reset the extension and IMC
+      // Reset the extension and IMC when the file URI changes
       if (imc) {
         imc.close();
         setImc(undefined);
@@ -236,7 +241,7 @@ export default function FileViewLoader({
       {usedExtension ? (
         <div className="relative h-full w-full">
           {!isExtensionLoaded && (
-            <div className="absolute left-0 top-0 h-full w-full">
+            <div className="absolute top-0 left-0 h-full w-full">
               <Loading />
             </div>
           )}
@@ -250,7 +255,7 @@ export default function FileViewLoader({
           )}
         </div>
       ) : hasExtension ? (
-        <div className="absolute left-0 top-0 h-full w-full">
+        <div className="absolute top-0 left-0 h-full w-full">
           <Loading />
         </div>
       ) : (