+++ npm-packages/shared-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pulse-editor/shared-utils
 
+## 0.1.1-alpha.7
+
+### Patch Changes
+
+- Simplify imc hooks
+
 ## 0.1.1-alpha.6
 
 ### Patch Changes